Extract scroll-position runtime code to separate .ts file
